+++ src/Controller/B2B/PackController.php
@@ -360,14 +360,19 @@ class PackController extends Controller
         $em = $this->getDoctrine()->getEntityManager();
 
         $media = $userPackMediaRepository->findBy(['name' => $request->get('name')]);
-
+        if(is_null($media))
+        {
+            return new JsonResponse(['success' => false]);
+        }
         $pack = $media[0]->getUserPack();
 
         $em->remove($media[0]);
 
         $em->flush();
+        return new JsonResponse(['success' => true]);
 
-        unlink('uploads/pack/'.$pack->getId().'/'.$request->get('name'));
+
+//        unlink('uploads/pack/'.$pack->getId().'/'.$request->get('name'));
 
         exit;
